Fix S&E borders being dropped from stitch
The canvas size was not large enough to fit the images (off by one tile in width and height).
We do download and paste these tiles into the image, but Pillow silently ignores images outside of canvas bounds.

Expand the canvas by one tile size in each direction to remedy this.